+++ src/keys.h
@@ -30,6 +30,9 @@ unsigned log_encoded_key(art::value_type key, bool start = true);
 template<typename UT>
 static art::node_ptr leaf_numeric_update(UT &l, const art::node_ptr &old, UT by) {
     const art::leaf *leaf = old.const_leaf();
+    if (leaf->is_compressed()) {
+        return nullptr;
+    }
     if (conversion::convert_value(l, leaf->get_value())) {
 
         auto& alloc = const_cast<alloc_pair&>(old.logical.get_ap<alloc_pair>());
@@ -49,6 +52,7 @@ static art::node_ptr leaf_numeric_update(UT &l, const art::node_ptr &old, UT by)
         ,  conversion::to_value(s)
         ,  leaf->expiry_ms()
         ,  leaf->is_volatile()
+        , leaf->is_compressed()
         );
     }
     return nullptr;